requirements.txt
pyyaml
nose
sphinx